requirements.txt
python-docx
python-dateutil
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
